requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
fastapi==0.115.13
filelock==3.18.0
fsspec==2025.5.1
googleapis-common-protos==1.70.0
grpcio==1.73.0
h11==0.16.0
hf-xet==1.1.4
huggingface-hub==0.33.0
idna==3.10
importlib_metadata==8.7.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==2.3.0
opentelemetry-api==1.34.1
opentelemetry-exporter-otlp==1.34.1
opentelemetry-exporter-otlp-proto-common==1.34.1
opentelemetry-exporter-otlp-proto-grpc==1.34.1
opentelemetry-exporter-otlp-proto-http==1.34.1
opentelemetry-instrumentation==0.55b1
opentelemetry-instrumentation-asgi==0.55b1
opentelemetry-instrumentation-fastapi==0.55b1
opentelemetry-instrumentation-logging==0.55b1
opentelemetry-proto==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
opentelemetry-util-http==0.55b1
packaging==25.0
protobuf==5.29.5
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
setuptools==80.9.0
sniffio==1.3.1
starlette==0.46.2
sympy==1.14.0
tokenizers==0.21.1
torch==2.7.1
tqdm==4.67.1
transformers==4.52.4
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
uvicorn==0.34.3
wrapt==1.17.2
zipp==3.23.0
